+++ src/StackTrace.php
@@ -17,8 +17,8 @@ use Composer\Semver\Semver;
  */
 class StackTrace implements StackTraceInterface
 {
-    const VERSION = '0.7.2';
-    const CONSTRAINTS_VERSION = '>=0.1.0 <0.8.0';
+    const VERSION = '0.8.0';
+    const CONSTRAINTS_VERSION = '>=0.1.0 <0.9.0';
 
     private $arrayStackTrace;