+++ src/VarDumpers/LightVarDumper.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Awesomite\StackTrace\VarDumpers;
-
-/**
- * @deprecated Class exists for backward compatibility, will be removed in new version
- */
-class LightVarDumper extends \Awesomite\VarDumper\LightVarDumper {}