+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.8.0
+
+* Deprecated classes `Awesomite\StackTrace\VarDumpers\*` have been removed
+
+## 0.7.2
+
+* Value of `StackTrace::VERSION` has been corrected
+
+## 0.7.1
+
+* Version of `awesomite/var-dumper` ha been updated to `^0.3.0`
+
 ## 0.7.0
 
 * Decreased size of serialized stack traced;